MNN

Documentation for MNN.

This project was created for Jugend Forscht 2024 (a German STEM competition).

Project Summary (German)

Wir wollen uns mit dem neuen, noch vergleichsweise wenig erforschten Bereich der mechanical neural networks, kurz MNNs, beschäftigen. MNNs sind programmierbare Materialien, welchen verschiedene Verhaltensweisen, wie zum Beispiel ein bestimmtes Verformungsverhalten, antrainiert werden können. Sie bestehen aus Massepunkten (genannt Neuronen), welche durch Federn miteinander verbunden werden. Ihr Verhalten ergibt sich durch die Steifheiten der Federn. Die grundlegende Annahme von MNNs ist, dass diese Federkonstanten in zukünftigen Materialien einzeln angepasst werden können. In Analogie zu künstlichen neuronalen Netzwerken wäre es dann prinzipiell möglich, durch eine geeignet gewählte Konfiguration an Federkonstanten verschiedene Verhaltensweisen auf externe Kräfte anzutrainieren. Während sich die bisherige Forschung auf die technische, physische Implementation dieser Netzwerke fokussiert hat, wollen wir das Trainingsverfahren optimieren. Dazu haben wir bereits eine Simulation eines MNNs umgesetzt, die bisher verwendeten Algorithmen (evolutionäres Lernen und Pattern Search) selbst implementiert, sowie mit neuen Parametern ausprobiert und verglichen. Dafür haben wir uns jedoch auf die Anwendung dieser Algorithmen in Simulationsrechnungen beschränkt. Die Ergebnisse sollten dennoch einen guten Startpunkt für reale MNNs bieten. Der von uns entwickelte Code ist die erste öffentlich verfügbare Implementation eines MNNs. Unsere Ergebnisse zeigen, dass MNNs mehrere komplexe Verhaltensweisen lernen können. Diese intelligenten Materialien eröffnen vielfältige zukünftige technologische Anwendungsmöglichkeiten.

Installation

Open the Julia REPL, then

julia> # type ] to open package manager

pkg> add https://github.com/Alexander-Reimer/Simulation-of-MNNs

This should install the MNN package and all dependencies automatically.

NixOS

When using NixOS, activating the development shell is necessary to make GLMakie work:

cd /path/to/this/project
nix-shell .

Usage

To use MNN.jl in you project, you should first import it with

using MNN

You can then create your first MNN (see MNN.get_user_behaviour for instructions on how to use the GUI):

# create a MNN  with 5 columns and 4 rows
net = Network(5, 4) 
# show the network
vis = Visualizer(net)
# create a behaviour to simulate using a GUI
b = get_user_behaviour(network::Network)
# simulate network with forces determined by b applied for 500 time steps
simulate!(net, Diff(500), b, vis = vis)

MNN.Network

MNN.Neuron

MNN.Spring

MNN.calc_ydist

MNN.get_user_behaviour

MNN.simulate!

MNN.Network — Type

mutable struct Network

Represents a MNN.

Fields

graph::MetaGraphsNext.MetaGraph: The graph representing the MNN (see MetaGraph.jl docs for more details). The edges are of type MNN.Spring, the node of type MNN.Neuron.

rows::Int: How many rows fixed columns have.

row_counts::Array{Int,1}: How many neurons are in each column. Access with column index (e.g. row_counts[1] == 5 –> the first column has 5 rows).

col_fixed::Array{Bool,1}: Whether the top and bottom neuron of a column are fixed (not movable). Access with column index like above, e.g. network.col_fixed[1] == true –> the first column is fixed, so get_neuron(network, 1, 1).movable == false.

columns::Int: How many columns the MNN has.

neuron_count::Int: How many neurons the MNN has.

xdist::Float64: How far apart the columns are initially on the x axis.

ydist::Float64: How far apart the neurons of a column are initially on the y axis.

start_positions::Dict{Int,Vector{Number}}: The initial positions of the neurons, accessed throught the neuron index. Example: network.start_positions[get_neuron_index(network, 1, 1)] should always return [0, 0].

positions::Array{Float64,2}: The current positions of the neurons. First index represents the space components, second one the neuron. So network.positions[:, 1] for position of first neuron, network.positions[2, 1] for just the y component of the first neuron.

velocities::Array{Float64,2}: The current velocities of the neurons. Same indexing as positions.

MNN.Neuron — Type

struct Neuron

Represents a node between springs. Position and velocity are stored in the MNN.Network.

Fields

movable::Bool: Is the neuron movable (not fixed to wall)?

MNN.Spring — Type

struct Spring

Represents an edge between neurons.

Fields

spring_constant::Float64: The spring constant.

length::Float64: The length of the spring at rest; this is constant and can't be changed after creation of the object.

MNN.calc_ydist — Method

calc_ydist(xdist)

Calculate necessary distance on y axis to create an equilateral triangle with given distance between columns on x axis.

MNN.get_user_behaviour — Method

get_user_behaviour(network::Network)

Get user input for behaviour.

MNN.simulate! — Method

simulate!(
network::Network,
sim::Simulation,
behaviour::Behaviour;
vis::Union{Visualizer,Nothing}=nothing,

)

TBW