Internal Documentation

Documentation for the internals of MNN.jl. These are subject to change.

Contents

Contents

Index

Public Interface

Index

MNN.Behaviour

MNN.House

MNN.Network

MNN.Neuron

MNN.Spring

MNN.calc_ydist

Public Interface

MNN.Behaviour — Type

abstract type Behaviour

Abstract supertype of all behaviours the network can be optimized for. Implemented behaviours:

TODO: Add implemented behaviours here.

MNN.House — Type

mutable struct House <: Network

Rough model of a house with a square layout and gravity. The purpose is to model earthquakes and analyse the effects on the "house" and effectiveness of training MNNs for earthquake / vibration protection.

MNN.Network — Type

abstract type Network

Abstract supertype of all mechanical neural networks. Implemented networks:

MNN.TestNetwork: Network described by Lee et al. (2022) with springs arrangd in equilateral triangles.

MNN.House: Rough model of a house with a square layout and gravity.

MNN.Neuron — Type

struct Neuron

Represents a node between springs. Position and velocity are stored in the MNN.Network.

Fields

movable::Bool: Is the neuron movable (not fixed to wall)?

MNN.Spring — Type

struct Spring

Represents an edge between neurons.

Fields

spring_constant::Float64: The spring constant.

length::Float64: The length of the spring at rest; this is constant and can't be changed after creation of the object.

MNN.calc_ydist — Method

calc_ydist(xdist)

Calculate necessary distance on y axis to create an equilateral triangle with given distance between columns on x axis.