Public Documentation

Documentation for MNN.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Contents

Index

Public Interface

Index

MNN.Deformation

MNN.Euler

MNN.Euler

MNN.Network

MNN.SecondOrderDiff

MNN.get_user_behaviour

MNN.simulate!

Public Interface

MNN.Deformation — Type

mutable struct Deformation <: Behaviour

A deformation behaviour for the MNN to learn.

Fields

goals: A dictionary with the goal positions of the neurons. Each key is a neuron index

with the value being a vector of the goal coordinates ([x, y]). If relative is true, then the coordinates will be interpreted as relative to the initial coordinates of the neurons, meaning as vector from the initial position, otherwise they will be interpreted as absolute coordinates.

relative: A boolean indicating whether the goal coordinates are relative or absolute.

modifiers: A dictionary with the modifiers for the neurons. Each key is a neuron index

with the value being a vector of the accelerations ([x, y]) which will be added to the velocity of the respective neuron continuously during the simulation,

Examples

MNN.Euler — Type

mutable struct Euler <: Simulation

MNN.Euler — Method

Euler(time::Number)

Create an Euler simulation with given time.

MNN.Network — Type

mutable struct Network

Represents a MNN.

Fields

graph::MetaGraphsNext.MetaGraph: The graph representing the MNN (see MetaGraph.jl docs for more details). The edges are of type MNN.Spring, the node of type MNN.Neuron.

rows::Int: How many rows fixed columns have.

row_counts::Array{Int,1}: How many neurons are in each column. Access with column index (e.g. row_counts[1] == 5 –> the first column has 5 rows).

col_fixed::Array{Bool,1}: Whether the top and bottom neuron of a column are fixed (not movable). Access with column index like above, e.g. network.col_fixed[1] == true –> the first column is fixed, so get_neuron(network, 1, 1).movable == false.

columns::Int: How many columns the MNN has.

neuron_count::Int: How many neurons the MNN has.

xdist::Float64: How far apart the columns are initially on the x axis.

ydist::Float64: How far apart the neurons of a column are initially on the y axis.

start_positions::Dict{Int,Vector{Number}}: The initial positions of the neurons, accessed throught the neuron index. Example: network.start_positions[get_neuron_index(network, 1, 1)] should always return [0, 0].

positions::Array{Float64,2}: The current positions of the neurons. First index represents the space components, second one the neuron. So network.positions[:, 1] for position of first neuron, network.positions[2, 1] for just the y component of the first neuron.

velocities::Array{Float64,2}: The current velocities of the neurons. Same indexing as positions.

MNN.SecondOrderDiff — Type

mutable struct SecondOrderDiff <: Simulation

MNN.get_user_behaviour — Method

get_user_behaviour(network::Network)

Get deformation behaviour using GUI. Left click on a neuron (first or last column) to select it, then release left mouse button and move mouse pointer to move the other end of the goal / force vector. Click left mouse button again to confirm.

If you have set all force and goal vectors you want, press space to finish. The function will then return a MNN.Deformation object.

MNN.simulate! — Method

simulate!(network::Network, sim::Simulation, behaviour::Behaviour; vis::Union{Visualizer,Nothing}=nothing)

Simulate network with given behaviour. Simulation method is determined by type of sim; currently MNN.SecondOrderDiff and MNN.Euler are implemented.

To implement your own, you need to define a struct that is a subtype of MNN.Simulation and has the field modifier::Function. Then overload the function MNN.simulate!(network::Network, sim::YourType; vis::Union{Visualizer, Nothing} = nothing).