Public Documentation

Documentation for MNN.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Contents

Index

Public Interface

Index

MNN.FirstOrderDiff

MNN.SecondOrderDiff

MNN.TestNetwork

MNN.get_user_behaviour

MNN.simulate!

Public Interface

MNN.FirstOrderDiff — Type

mutable struct FirstOrderDiff <: Simulation

MNN.SecondOrderDiff — Type

mutable struct SecondOrderDiff <: Simulation

MNN.TestNetwork — Type

mutable struct TestNetwork <: Network

Represents a MNN as described by Lee et al. (2022) with springs arranged in equilateral triangles. The network is initialized with

TestNetwork(columns::Int, rows::Int)

MNN.get_user_behaviour — Method

get_user_behaviour(network::Network)

Get deformation behaviour using GUI. Left click on a neuron (first or last column) to select it, then release left mouse button and move mouse pointer to move the other end of the goal / force vector. Click left mouse button again to confirm.

If you have set all force and goal vectors you want, press space to finish. The function will then return a MNN.Behaviour object.

MNN.simulate! — Method

simulate!(network::Network, sim::Simulation, behaviour::Behaviour; vis::Union{Visualizer,Nothing}=nothing)

Simulate network with given behaviour. Simulation method is determined by type of sim. TODO: Add implemented simulation methods here. Refer to page for own simulation method.